.history/requirements_20250516195619.txt
customtkinter==5.2.2
Pillow==10.2.0
pandas==2.2.1
openpyxl==3.1.2
python-docx==0.8.11 